Move classes to the shared project.
