+++ test/shared/src/main/java/org/elasticsearch/hadoop/serialization/AbstractValueReaderTest.java
