+++ test/shared/src/main/java/org/elasticsearch/hadoop/serialization/JdkValueReaderTest.java
@@ -23,9 +23,11 @@ import org.elasticsearch.hadoop.serialization.builder.ValueReader;
 import org.elasticsearch.hadoop.util.DateUtils;
 import org.junit.Test;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertArrayEquals;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertThat;
 
-import static org.hamcrest.CoreMatchers.*;
+import static org.hamcrest.CoreMatchers.nullValue;
 
 
 public class JdkValueReaderTest extends AbstractValueReaderTest {
