+++ test/shared/src/main/java/org/elasticsearch/hadoop/serialization/JsonUtils.java
