+++ test/shared/src/main/java/org/elasticsearch/hadoop/serialization/builder/AbstractExtendedBooleanValueReaderTest.java
