+++ test/shared/src/main/java/org/elasticsearch/hadoop/util/TestSettings.java
